Filter topics using naming convention vs module db
Topic re-publishing doesn't have to know or care about database state.
This commit makes topic filtering more loosely coupled. By inferring the
sensor endpoints via topic name, we no longer have to care _who_ is
publishing that topic, or _how_. We just have to care that the sensor
topic exists.

As a fringe benefit, this also reduces the # of lines.