Remove ROS dependency for building docs
Updated the doc/conf.py file to parse the xml data in package.xml
for the version information instead of relying on the catkin_pkg
import and ROS dependency.

Previously, unless ROS was installed on the system running the doc
build, the ``make html`` command would fail. Now ROS is no longer
required to build the docs.

References Issue #78.