Add workaround for water level and EWMA filter
This is a short-term fix for preventing boolean values from
being filtered by the EWMA filter.

Explanation: right now all topics under `/environment/<id>` are
float64 type, with Boolean being 1 or 0. This was judged to be a
simpler architecture at the time. Values from sensors may be any
type, but are coerced to Float64. The same is true for actuators.
However, this assumption breaks down for filtering boolean values,
since the EWMA will produce fractional numbers that don't coerce
correctly back to boolean.

In future, we should change the architecture of the system to support
standard ros types under `/environment/<id>`.