Read env vars from params explicitly
This is to fix a potential problem with key ordering of dicts.

Rather than reading a dict into a list, we let the consumer decide how
to read the dict.

A from_dict staticmethod is added to VariableInfo to give consumers an
easier way to convert from YAML to class instance.