Fix Dockerfile
Currently the Dockerfile is broken due to changes to the install scripts.
The Dockerfile now mostly relies on `scripts/install_dev`, and due to it's
dependency on resin's raspbian image, a couple of bits had to be added.

- Create gpio group
- Give the user pi access to video group

I also had to make some changes to the installation scripts, given the current
state of the raspbian image we use for the docker build:

Commented out the curl request that configs compaction on the db, because during
the build process for the brain the db is not expected to be up. I'd say it
should be executed at runtime in some kind of initialization task.

Changed virtualenv to be isntalled via pip, for some reason installing it via
apt-get wasn't working on my pi