Move out couchdb related stuff from init_dev
The Docker build relies on install_dev, which executes init_dev. This has issues
with the docker build, because couchdb is expected to run in a separate
container.

I moved out the compaction setting to the db config in db_bootstrap/db_config.py
Also moved out the installation of couchdb to it's own install_db script.